+++ src/View/InputView.js
@@ -1,6 +1,6 @@
 const { Console } = require('@woowacourse/mission-utils');
 const OutputView = require('./OutputView');
-const Validation = require('./Validation');
+const Validation = require('../utils/Validation');
 
 const InputView = {
   CAR_NAMES: '경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)\n',
